:loud_sound:添加日志;:white_check_mark:下载出错时返回消息给浏览器扩展

+++ extension.js
@@ -108,6 +108,8 @@ function parseExtensions() {
 
 function installDependencies() {
   for (let name in installPackages) {
+    const stat = fs.statSync(path.resolve(__dirname, './extensions/node_modules/' + name))
+    if (stat.isFile() || stat.isDirectory()) continue
     console.info(`install ${name}@${installPackages[name]}`)
     runCommand('npm install --prefix extensions ' + name + '@' + installPackages[name])
   }
